Move error handling to its own class
